requirements.txt
Django==1.10.2
django-appconf==1.0.2
django-compressor==2.1
django-filter==0.15.3
djangorestframework==3.4.7
djangorestframework-gis==0.10.1
djoser==0.5.1
easy-thumbnails==2.3
Pillow==3.4.1
psycopg2==2.6.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
unicode-slugify==0.1.3
Unidecode==0.4.19
xlrd==1.0.0
